Update E2E tests.
